Split connectors cli
