+++ connectors/src/connectors/slack/lib/cli.ts
@@ -1,5 +1,17 @@
-import { launchSlackSyncWorkflow } from "@connectors/connectors/slack/temporal/client";
+import type {
+  AdminSuccessResponseType,
+  SlackCommandType,
+} from "@dust-tt/types";
+
+import {
+  launchSlackSyncOneThreadWorkflow,
+  launchSlackSyncWorkflow,
+} from "@connectors/connectors/slack/temporal/client";
+import { throwOnError } from "@connectors/lib/cli";
 import { SlackChannel } from "@connectors/lib/models/slack";
+import { default as topLogger } from "@connectors/logger/logger";
+import { SlackConfigurationResource } from "@connectors/resources/slack_configuration_resource";
+import { ConnectorModel } from "@connectors/resources/storage/models/connector_model";
 
 export async function maybeLaunchSlackSyncWorkflowForChannelId(
   connectorId: number,
@@ -19,3 +31,170 @@ export async function maybeLaunchSlackSyncWorkflowForChannelId(
 
   return launchSlackSyncWorkflow(connectorId, null, [slackChannelId]);
 }
+
+export const slack = async ({
+  command,
+  args,
+}: SlackCommandType): Promise<AdminSuccessResponseType> => {
+  const logger = topLogger.child({ majorCommand: "slack", command, args });
+  switch (command) {
+    case "enable-bot": {
+      if (!args.wId) {
+        throw new Error("Missing --wId argument");
+      }
+      const connector = await ConnectorModel.findOne({
+        where: {
+          workspaceId: `${args.wId}`,
+          type: "slack",
+        },
+      });
+      if (!connector) {
+        throw new Error(`Could not find connector for workspace ${args.wId}`);
+      }
+      const slackConfig = await SlackConfigurationResource.fetchByConnectorId(
+        connector.id
+      );
+      if (!slackConfig) {
+        throw new Error(
+          `Could not find slack configuration for connector ${connector.id}`
+        );
+      }
+
+      const res = await slackConfig.enableBot();
+      if (res.isErr()) {
+        throw res.error;
+      }
+      return { success: true };
+    }
+
+    case "sync-channel": {
+      const { channelId, wId } = args;
+
+      if (!wId) {
+        throw new Error("Missing --wId argument");
+      }
+      if (!channelId) {
+        throw new Error("Missing --channelId argument");
+      }
+
+      const connector = await ConnectorModel.findOne({
+        where: {
+          workspaceId: wId,
+          type: "slack",
+        },
+      });
+      if (!connector) {
+        throw new Error(`Could not find connector for workspace ${wId}`);
+      }
+
+      await throwOnError(
+        maybeLaunchSlackSyncWorkflowForChannelId(connector.id, channelId)
+      );
+
+      return { success: true };
+    }
+
+    case "sync-thread": {
+      if (!args.wId) {
+        throw new Error("Missing --wId argument");
+      }
+      if (!args.threadId) {
+        throw new Error("Missing --threadId argument");
+      }
+      if (!args.channelId) {
+        throw new Error("Missing --channelId argument");
+      }
+      const connector = await ConnectorModel.findOne({
+        where: {
+          workspaceId: `${args.wId}`,
+          type: "slack",
+        },
+      });
+      if (!connector) {
+        throw new Error(`Could not find connector for workspace ${args.wId}`);
+      }
+      await throwOnError(
+        launchSlackSyncOneThreadWorkflow(
+          connector.id,
+          args.channelId,
+          args.threadId
+        )
+      );
+
+      return { success: true };
+    }
+
+    case "whitelist-domains": {
+      const { wId, whitelistedDomains } = args;
+      if (!wId) {
+        throw new Error("Missing --wId argument");
+      }
+      if (!whitelistedDomains) {
+        throw new Error("Missing --whitelistedDomains argument");
+      }
+
+      const connector = await ConnectorModel.findOne({
+        where: {
+          workspaceId: `${args.wId}`,
+          type: "slack",
+        },
+      });
+
+      if (!connector) {
+        throw new Error(`Could not find connector for workspace ${args.wId}`);
+      }
+
+      const whitelistedDomainsArray = whitelistedDomains.split(",");
+      // TODO(2024-01-10 flav) Add domain validation.
+      logger.info(
+        `[Admin] Whitelisting following domains for slack:\n- ${whitelistedDomainsArray.join(
+          "\n-"
+        )}`
+      );
+
+      const slackConfig = await SlackConfigurationResource.fetchByConnectorId(
+        connector.id
+      );
+      if (slackConfig) {
+        await slackConfig.setWhitelistedDomains(whitelistedDomainsArray);
+      }
+
+      return { success: true };
+    }
+
+    case "whitelist-bot": {
+      const { wId, botName } = args;
+      if (!wId) {
+        throw new Error("Missing --wId argument");
+      }
+      if (!botName) {
+        throw new Error("Missing --botName argument");
+      }
+
+      const connector = await ConnectorModel.findOne({
+        where: {
+          workspaceId: `${args.wId}`,
+          type: "slack",
+        },
+      });
+
+      if (!connector) {
+        throw new Error(`Could not find connector for workspace ${args.wId}`);
+      }
+
+      logger.info(`[Admin] Whitelisting following bot for slack: ${botName}`);
+
+      const slackConfig = await SlackConfigurationResource.fetchByConnectorId(
+        connector.id
+      );
+      if (slackConfig) {
+        await slackConfig.whitelistBot(botName);
+      }
+
+      return { success: true };
+    }
+
+    default:
+      throw new Error("Unknown slack command: " + command);
+  }
+};
